+++ docs/RELEASE-NOTES.rst
@@ -1,6 +1,20 @@
 Release Notes for Container Ingress Services for Kubernetes & OpenShift
 =======================================================================
 
+
+2.10.1
+-------------
+Bug Fixes
+````````````
+* Fix to monitor NGINX+ service changes
+* :issues:`2582` Fix issue with inconsistent pool names for VS
+* :issues:`2596` Fix invalid property name with serviceAddress
+* :issues:`2570` Fix for TLSProfile doesn't get updated when K8s secret changes
+* :issues:`2394` Fix to set ingress https monitor send string
+* :issues:`2549` Fix trafficGroup regex
+* :issues:`2492` Fix for shared pool not working in nodePort mode
+
+
 2.10.0
 -------------